Use forms module, adjust struct types
