+++ src/forms/mod.rs
@@ -1,3 +1,7 @@
 mod rating;
 
+pub(crate) mod stack;
+
 pub use rating::*;
+
+pub use stack::*;
