+++ src/models/mod.rs
@@ -3,3 +3,5 @@ pub mod stack;
 pub mod user;
 
 pub use rating::*;
+
+pub use stack::*;
